+++ NEWS.md
@@ -1,3 +1,12 @@
+# fingertipsR 1.1.0 (2025-10-10)
+
+* Updated the description of the nearest_neighbours function, and removed the measure variable that was previously deprecated. 
+* Added checks to handle cases where no data is available for the specified combination of parameters in `fingertips_data()`.
+* When no data is available and `rank = TRUE`, the function now returns an empty data frame with `Polarity`, `Rank`, and `AreaValuesCount` columns set to `NA`.
+* Improved error message to inform users that no data is available and an empty data frame is returned.
+* Updated examples. 
+* Updated area_types() function to use the short name in the API rather than the long name 
+
 # fingertipsR 1.0.13 (2025-05-29)
 
 * no change for users. Package tests updated.